+++ gameplay/elements/unit/unit.gd
@@ -9,7 +9,7 @@ var moved: bool = false
 var used_ability: bool = false
 
 func setup():
-	#%Sprite2D.texture = character.sprite
+	%Sprite2D.texture = character.sprite
 	position = GameplayController.cell_to_position(cell)
 	print("Hello!")