+++ gameplay/tile/tile.gd
@@ -1,0 +1,7 @@
+class_name Tile
+
+var unit: Unit
+var type: Enums.RoomTargetType
+#var effect
+func _init(_type) -> void:
+	type = _type